Switch to gradle build system.
