+++ buildartifacts-auth-common/build.gradle
@@ -1,0 +1,5 @@
+dependencies {
+    compile group: 'com.google.http-client', name: 'google-http-client', version:'1.30.2'
+    compile group: 'com.google.auth', name: 'google-auth-library-oauth2-http', version:'0.16.2'
+    testCompile group: 'junit', name: 'junit', version: '4.12'
+}
