+++ buildartifacts-auth-common/src/main/java/com/google/cloud/buildartifacts/auth/GcloudCredentials.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.cloud.buildartifacts.wagon;
+package com.google.cloud.buildartifacts.auth;
 
 import com.google.api.client.json.JsonFactory;
 import com.google.api.client.json.jackson2.JacksonFactory;
