+++ buildartifacts-maven-wagon/build.gradle
@@ -1,0 +1,9 @@
+dependencies {
+    compile group: 'org.apache.maven.wagon', name: 'wagon-http-shared', version: '3.3.3'
+    compile group: 'org.apache.maven', name: 'maven-plugin-api', version: '3.6.1'
+    compile group: 'com.google.http-client', name: 'google-http-client', version: '1.30.2'
+    compile group: 'com.google.auth', name: 'google-auth-library-oauth2-http', version: '0.16.2'
+    testCompile group: 'org.apache.maven.wagon', name: 'wagon-provider-test', version: '3.3.3'
+    testCompile group: 'junit', name: 'junit', version: '4.12'
+    compileOnly group: 'org.apache.maven.wagon', name: 'wagon-provider-api', version: '3.3.3'
+}
