+++ buildartifacts-maven-wagon/src/main/resources/META-INF/plexus/components.xml
