+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Tue Jul 30 15:34:22 PDT 2019
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-5.5.1-all.zip
